:art::zap::lipstick: Match height of products on index page
Products page wasn't properly aligned due to different ratios on the images and different lengths on descriptions. Corrected this the proper way as opposed to just using same length descriptions and images.

A lot of the view was refactored as well as a third-party jQuery script being used. Learning documentation took forever, am using the API to control the height, not classnames